backend/requirements.txt
asgiref==3.11.0
dj-rest-auth==7.0.1
Django==5.2.9
django-allauth==65.13.1
django-cors-headers==4.9.0
djangorestframework==3.16.1
pillow==12.0.0
sqlparse==0.5.4
tzdata==2025.3
requests==2.32.5